+++ Tests/PerseusGeoLocationKitTests/AskForCurrentLocationTests.swift
@@ -37,11 +37,11 @@ extension PerseusLocationDealerTests {
 
         // arrange
 
-#if os(iOS)
+        #if os(iOS)
         MockLocationManager.status = .authorizedAlways
-#elseif os(macOS)
+        #elseif os(macOS)
         MockLocationManager.status = .authorized
-#endif
+        #endif
         MockLocationManager.isLocationServiceEnabled = true
 
         // act
@@ -55,22 +55,22 @@ extension PerseusLocationDealerTests {
         XCTAssertTrue(sut.order == .currentLocation)
         XCTAssertEqual(sut.locationManager.desiredAccuracy, APPROPRIATE_ACCURACY.rawValue)
 
-#if os(iOS)
+        #if os(iOS)
         mockLM.verify_requestLocation_CalledOnce()
-#elseif os(macOS)
+        #elseif os(macOS)
         mockLM.verify_startUpdatingLocation_CalledOnce()
-#endif
+        #endif
     }
 
     func test_askForCurrentLocation_should_set_the_accuracy() {
 
         // arrange
 
-#if os(iOS)
+        #if os(iOS)
         MockLocationManager.status = .authorizedAlways
-#elseif os(macOS)
+        #elseif os(macOS)
         MockLocationManager.status = .authorized
-#endif
+        #endif
         MockLocationManager.isLocationServiceEnabled = true
 
         // act
